restructured code, modified documentation
