+++ doc/implementation.rst
@@ -1,24 +1,34 @@
-**************
 Implementation
-**************
+===============
 
-This is where all the details regarding the implementation, 
-reading/writing of files, and process flow will be described.
+You can run the code by typing::
 
-Glossary
-========
-Full Load Hours
----------------
+	$ python Master.py
+	
+``Master.py`` includes the main functions for the code, which are explained in the following sections.
 
-Quantiles
----------
+.. toctree::
+   :maxdepth: 3
+   
+   source/Master
 
-...
-***
+Helping functions for the models are included in ``model_functions.py``.
 
+.. toctree::
+   :maxdepth: 3
+   
+   source/model_functions
 
-Helping functions
-*****************
+Helping functions for data formatting, filtering, reshaping, etc. are included in ``data_functions.py``.
 
-Regression
-**********
+.. toctree::
+   :maxdepth: 3
+   
+   source/data_functions
+   
+Utility functions as well as imported libraries are included in ``util.py``.
+
+.. toctree::
+   :maxdepth: 3
+   
+   source/util
