+++ doc/source/config.rst
@@ -1,0 +1,7 @@
+config module
+=============
+
+.. automodule:: config
+   :members:
+   :undoc-members:
+   :show-inheritance:
