+++ doc/source/data_functions.rst
@@ -1,0 +1,7 @@
+data\_functions module
+======================
+
+.. automodule:: data_functions
+   :members:
+   :undoc-members:
+   :show-inheritance:
