+++ doc/source/model_functions.rst
@@ -1,0 +1,5 @@
+model\_functions module
+=======================
+
+.. automodule:: model_functions
+   :members:
