+++ doc/source/util.rst
@@ -1,0 +1,7 @@
+util module
+===========
+
+.. automodule:: util
+   :members:
+   :undoc-members:
+   :show-inheritance:
